requirements.txt
numpy==1.23.2
pandas==1.3.3
py4j==0.10.9.5
pyarrow==5.0.0
pyspark==3.3.0
python-dateutil==2.8.1
scipy==1.9.1
six==1.15.0
coverage
pyyaml